refactor(desktop): extract composer seed helpers into composer-seed module
