+++ desktop/src/renderer/src/features/coding-agents/AgentComposer.tsx
@@ -12,63 +12,12 @@ import { useCodingAgentWorkspace } from "../../stores/coding-agent-workspace";
 import { useProviderPreferences } from "../settings/provider-preferences";
 import { AgentWorkspaceSection as Section } from "./AgentWorkspaceSection";
 import { capabilityEnabled } from "./capabilities";
+import { clearComposerLaunchContext, mergeComposerSeed, type ComposerSeed } from "./composer-seed";
 
-export type ComposerSeed = {
-  seedId: number;
-  draft: AgentThreadComposerDraft;
-};
-
-export function mergeAttachments(
-  current: AgentThreadComposerDraft["attachments"],
-  seeded: AgentThreadComposerDraft["attachments"],
-): AgentThreadComposerDraft["attachments"] {
-  const currentAttachments = current ?? [];
-  const seededById = new Map((seeded ?? []).map((attachment) => [attachment.id, attachment]));
-  const merged = currentAttachments.map((attachment) => seededById.get(attachment.id) ?? attachment);
-  const seen = new Set(merged.map((attachment) => attachment.id));
-  for (const attachment of seeded ?? []) {
-    if (seen.has(attachment.id) || merged.length >= 8) continue;
-    seen.add(attachment.id);
-    merged.push(attachment);
-  }
-  return merged.length ? merged : undefined;
-}
-
-export function mergeComposerSeed(current: AgentThreadComposerDraft, seeded: AgentThreadComposerDraft): AgentThreadComposerDraft {
-  const currentPrompt = current.prompt.trim();
-  const seededPrompt = seeded.prompt.trim();
-  const attachments = mergeAttachments(current.attachments, seeded.attachments);
-  const missingRequiredReference = (seeded.attachments ?? [])
-    .some((attachment) => attachment.kind === "structured_ref"
-      && !attachments?.some((candidate) => candidate.id === attachment.id));
-  if (missingRequiredReference) return current;
-
-  return {
-    ...seeded,
-    providerId: current.providerId ?? seeded.providerId,
-    mode: current.mode ?? seeded.mode,
-    approvalPolicy: current.approvalPolicy ?? seeded.approvalPolicy,
-    sandboxMode: current.sandboxMode ?? seeded.sandboxMode,
-    prompt: !seededPrompt
-      ? current.prompt
-      : currentPrompt && currentPrompt !== seededPrompt
-        ? `${current.prompt.trimEnd()}\n\n---\n\n${seeded.prompt}`
-        : seeded.prompt,
-    attachments,
-  };
-}
-
-export function clearComposerLaunchContext(current: AgentThreadComposerDraft): AgentThreadComposerDraft {
-  const attachments = current.attachments?.filter((attachment) => attachment.kind !== "structured_ref");
-  return {
-    ...current,
-    projectId: undefined,
-    taskId: undefined,
-    terminalSessionId: undefined,
-    worktreeId: undefined,
-    attachments: attachments?.length ? attachments : undefined,
-  };
-}
+// The seed helpers live in composer-seed.ts so the draft-chat pane and this
+// legacy panel composer share one implementation; re-export for existing
+// imports (ProjectChatsView re-exports them for tests).
+export { mergeAttachments, mergeComposerSeed, clearComposerLaunchContext, type ComposerSeed } from "./composer-seed";
 
 function hasComposerContent(current: AgentThreadComposerDraft): boolean {
   return current.prompt.trim().length > 0
