+++ desktop/src/renderer/src/features/coding-agents/composer-seed.ts
@@ -1,0 +1,62 @@
+import type { AgentThreadComposerDraft } from "@matrix-os/contracts";
+
+// Seed payloads for the new-chat draft: one-shot instructions merged into the
+// visible draft composer (review-hunk follow-ups, suggestion chips, typed
+// first characters, per-task targeting). `seedId` lets a repeated identical
+// seed re-apply after the draft was reset.
+export type ComposerSeed = {
+  seedId: number;
+  draft: AgentThreadComposerDraft;
+};
+
+export function mergeAttachments(
+  current: AgentThreadComposerDraft["attachments"],
+  seeded: AgentThreadComposerDraft["attachments"],
+): AgentThreadComposerDraft["attachments"] {
+  const currentAttachments = current ?? [];
+  const seededById = new Map((seeded ?? []).map((attachment) => [attachment.id, attachment]));
+  const merged = currentAttachments.map((attachment) => seededById.get(attachment.id) ?? attachment);
+  const seen = new Set(merged.map((attachment) => attachment.id));
+  for (const attachment of seeded ?? []) {
+    if (seen.has(attachment.id) || merged.length >= 8) continue;
+    seen.add(attachment.id);
+    merged.push(attachment);
+  }
+  return merged.length ? merged : undefined;
+}
+
+export function mergeComposerSeed(current: AgentThreadComposerDraft, seeded: AgentThreadComposerDraft): AgentThreadComposerDraft {
+  const currentPrompt = current.prompt.trim();
+  const seededPrompt = seeded.prompt.trim();
+  const attachments = mergeAttachments(current.attachments, seeded.attachments);
+  const missingRequiredReference = (seeded.attachments ?? [])
+    .some((attachment) => attachment.kind === "structured_ref"
+      && !attachments?.some((candidate) => candidate.id === attachment.id));
+  if (missingRequiredReference) return current;
+
+  return {
+    ...seeded,
+    providerId: current.providerId ?? seeded.providerId,
+    mode: current.mode ?? seeded.mode,
+    approvalPolicy: current.approvalPolicy ?? seeded.approvalPolicy,
+    sandboxMode: current.sandboxMode ?? seeded.sandboxMode,
+    prompt: !seededPrompt
+      ? current.prompt
+      : currentPrompt && currentPrompt !== seededPrompt
+        ? `${current.prompt.trimEnd()}\n\n---\n\n${seeded.prompt}`
+        : seeded.prompt,
+    attachments,
+  };
+}
+
+export function clearComposerLaunchContext(current: AgentThreadComposerDraft): AgentThreadComposerDraft {
+  const attachments = current.attachments?.filter((attachment) => attachment.kind !== "structured_ref");
+  return {
+    ...current,
+    projectId: undefined,
+    taskId: undefined,
+    terminalSessionId: undefined,
+    worktreeId: undefined,
+    attachments: attachments?.length ? attachments : undefined,
+  };
+}
